Roadmap truth pass + R39-DECOMP-VIEWER (17)
TWO STALE CLAIMS CORRECTED, both found by testing the entry against the tree.

R38-SYNC-2D3D's stated defect is fixed end to end. The entry said the pipeline
"discards element identity at bake time" and that "nothing in a plan can name
what it draws". `_bake_uncached` returns (guid, ifc_class, mesh) and its own
docstring credits R38-PLAN-IDENTITY; `cut_baked_guided` emits (guid, class,
polyline) with a PRODUCTION caller in the plan renderer; the SVG carries
data-guid and planPane.ts selects on it. The entry described two functions
accurately and drew the wrong conclusion because it never looked for a third.
What actually remains: it claims three open children and names none.

CLAUDE.md's viewer numbers were both stale — "twenty-eight commits" and "3,444
lines" are now 67 and 2,570. Unlike the Node and Python drifts that file already
records, this one moved in the direction that STRENGTHENS its argument, which is
the hardest kind to notice: a number that decays toward the conclusion it
supports never looks wrong.

R39-DECOMP-VIEWER (17) — field verification out of app.ts (2,571 -> 2,508).

app.ts is not a class, so REL-4's "grep the this. refs first" rule has no this.
to grep: the file is ONE 2,445-line function and everything in it is a closure.
The equivalent is how many SIBLING closures a candidate captures, and over all
fourteen candidates >=25 lines exactly ONE captured zero. buildToolsPanel
captures 14, handleKey 12, selectByGuids 6 — every other move would have been the
callback bag REL-4 warns about. Four of five free variables already travelled on
the typed ViewerCtx, so the deps object is that context narrowed.

THE NARRATIVE-CHAIN GATE REFUSED THE FIRST RATCHET ENTRY, CORRECTLY. The previous
entry ended at 2,571 and the file measured 2,570 — one line had left with no
slice recording it, the drift the roadmap cell already documents from another
lane. The entry now runs 2,571 -> 2,508 and names the stray rather than starting
at a number nobody can reproduce. Ratchet mutation-checked at 2,507.

2072 web tests, tsc + eslint + build clean, 8 structural gates green.
