+++ services/api/test_file_sizes.py
@@ -124,7 +124,7 @@ PER_FILE = {
     #: construction, and the three slices' friction was spent silently. Pinned at 2_570, the file's
     #: EXACT measured size — ⑯ set 2_571 against a file that already measured 2,570, which is the
     #: off-by-one MAX_SLACK is deliberately loose enough to tolerate.
-    "apps/web/src/viewer/app.ts": 2_570,   # detailing -> tools/detailingSection.ts (2_630 -> 2_571).   # content + family library -> tools/contentLibrarySection.ts (2_757 -> 2_630).   # interactive annotation -> tools/annotationSection.ts (2_865 -> 2_757). The file was AT this pin with zero headroom, which is why the roadmap row calling it "97%, ~136 lines" was corrected in the same pass.   # CAD command line -> cadBar.ts, which also gained the R29 prompt loop (2_885 -> 2_865). The 39-line inline block became a 10-line mount call; the ratchet is why the interactive mode went into a new module instead of on top of app.ts.   # clash rail -> tools/clashPanel.ts (2_944 -> 2_885)
+    "apps/web/src/viewer/app.ts": 2_508,   # R39-DECOMP-VIEWER (17): field verification -> tools/verifySection.ts (2_571 -> 2_508). The chain starts at 2_571 and the file measured 2_570 when this slice began: ONE line left with no slice recording it, which is the drift this row's roadmap cell already documents from an unrelated lane committing to the same shared file. This slice removed 62 of the 63; the 63rd is that stray, absorbed here rather than hidden by starting the entry at a number nobody can reproduce. The gate is what refused the discontinuity. app.ts is ONE 2,445-line function, so REL-4's "grep the this. refs first" rule has no this. to grep; the equivalent is how many SIBLING closures a candidate captures, and of fourteen candidates >=25 lines exactly ONE captured zero. buildToolsPanel captures 14, handleKey 12, selectByGuids 6 — every other move would have been a callback bag. Four of renderVerify's five free variables already travelled on the typed ViewerCtx, so the deps object is that context narrowed, not one invented to make the move possible.   # detailing -> tools/detailingSection.ts (2_630 -> 2_571).   # content + family library -> tools/contentLibrarySection.ts (2_757 -> 2_630).   # interactive annotation -> tools/annotationSection.ts (2_865 -> 2_757). The file was AT this pin with zero headroom, which is why the roadmap row calling it "97%, ~136 lines" was corrected in the same pass.   # CAD command line -> cadBar.ts, which also gained the R29 prompt loop (2_885 -> 2_865). The 39-line inline block became a 10-line mount call; the ratchet is why the interactive mode went into a new module instead of on top of app.ts.   # clash rail -> tools/clashPanel.ts (2_944 -> 2_885)
     # Pinned at its EXACT measured size, not above it. qaSection.ts became the file every reach fix
     # lands in and reached 1,373 lines while unpinned - the same accumulation app.ts and client.ts
     # already have entries for. Pinned before it needs splitting rather than after: a ratchet added
